requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# Optimization libraries
scipy>=1.11.0
cvxpy>=1.4.0
pulp>=2.7.0
numpy>=1.24.0

# PDF processing
PyPDF2>=3.0.0

# Image processing
Pillow>=10.0.0

# Utilities
pydantic>=2.0.0
pyyaml>=6.0
